+++ packages/riverpod/lib/src/framework/container.dart
@@ -268,7 +268,7 @@ class ProviderContainer implements Node {
     );
   }
 
-  /// {@template riverpod.invalidate}
+  /// {@macro riverpod.invalidate}
   void invalidate(ProviderOrFamily provider) {
     if (provider is ProviderBase) {
       final reader = _getStateReader(provider._origin);
@@ -287,7 +287,7 @@ class ProviderContainer implements Node {
     }
   }
 
-  /// {@template riverpod.refresh}
+  /// {@macro riverpod.refresh}
   State refresh<State>(Refreshable<State> provider) {
     invalidate(provider._origin);
     return read(provider);