+++ packages/riverpod_generator/lib/src/templates/family.dart
@@ -1,5 +1,4 @@
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/element/element.dart';
 import 'package:collection/collection.dart';
 import 'package:riverpod_analyzer_utils/riverpod_analyzer_utils.dart';
 
@@ -74,17 +73,17 @@ class FamilyTemplate extends Template {
     var providerType = '${leading}Provider';
     var refType = '${leading}ProviderRef';
     var elementType = '${leading}ProviderElement';
-    var createdType = provider.createdType.toString();
+    var createdType = provider.createdTypeDisplayString;
 
-    final returnType = provider.createdType;
-    if (!returnType.isRaw) {
+    final returnType = provider.createdTypeNode?.type;
+    if (returnType != null && !returnType.isRaw) {
       if (returnType.isDartAsyncFutureOr || returnType.isDartAsyncFuture) {
         providerType = '${leading}FutureProvider';
         refType = '${leading}FutureProviderRef';
         elementType = '${leading}FutureProviderElement';
         // Always use FutureOr<T> in overrideWith as return value
         // or otherwise we get a compilation error.
-        createdType = 'FutureOr<${provider.valueType}>';
+        createdType = 'FutureOr<${provider.valueTypeDisplayString}>';
       } else if (returnType.isDartAsyncStream) {
         providerType = '${leading}StreamProvider';
         refType = '${leading}StreamProviderRef';
@@ -92,14 +91,12 @@ class FamilyTemplate extends Template {
       }
     }
 
-    final parameters = provider
-        .node.functionExpression.parameters!.parameterElements
-        .whereNotNull()
+    final parameters = provider.node.functionExpression.parameters!.parameters
         .skip(1)
         .toList();
 
     final parametersPassThrough = buildParamInvocationQuery({
-      for (final parameter in parameters) parameter: parameter.name,
+      for (final parameter in parameters) parameter: parameter.name!.lexeme,
     });
 
     final typeParameters = provider.node.functionExpression.typeParameters;
@@ -113,7 +110,7 @@ class FamilyTemplate extends Template {
       hashFn: hashFn,
       elementType: elementType,
       refType: refType,
-      providerGenerics: '<${provider.valueType}>',
+      providerGenerics: '<${provider.valueTypeDisplayString}>',
       providerCreate:
           '(ref) => ${provider.name}$typeParametersUsage(ref as ${provider._refImplName}$typeParametersUsage, $parametersPassThrough)',
       providerType: providerType,
@@ -157,8 +154,8 @@ ${parameters.map((e) => '        ${e.name}: ${e.name},\n').join()}
     var notifierBaseType = 'Buildless${leading}Notifier';
     var elementType = '${leading}NotifierProviderElement';
 
-    final returnType = provider.createdType;
-    if (!returnType.isRaw) {
+    final returnType = provider.createdTypeNode?.type;
+    if (returnType != null && !returnType.isRaw) {
       if (returnType.isDartAsyncFutureOr || returnType.isDartAsyncFuture) {
         providerType = '${leading}AsyncNotifierProviderImpl';
         refType = '${leading}AsyncNotifierProviderRef';
@@ -172,15 +169,14 @@ ${parameters.map((e) => '        ${e.name}: ${e.name},\n').join()}
       }
     }
 
-    final parameters = provider.buildMethod.parameters!.parameterElements
-        .whereNotNull()
-        .toList();
+    final parameters =
+        provider.buildMethod.parameters!.parameters.whereNotNull().toList();
     final parameterDefinition = buildParamDefinitionQuery(parameters);
     final cascadePropertyInit =
         parameters.map((e) => '..${e.name} = ${e.name}').join('\n');
 
     final parametersPassThrough = buildParamInvocationQuery({
-      for (final parameter in parameters) parameter: parameter.name,
+      for (final parameter in parameters) parameter: parameter.name!.lexeme,
     });
 
     final typeParameters = provider.node.typeParameters;
@@ -197,7 +193,7 @@ ${parameters.map((e) => '        ${e.name}: ${e.name},\n').join()}
       elementType: elementType,
       refType: refType,
       providerGenerics:
-          '<${provider.name}$typeParametersUsage, ${provider.valueType}>',
+          '<${provider.name}$typeParametersUsage, ${provider.valueTypeDisplayString}>',
       providerType: providerType,
       providerCreate: parameters.isEmpty
           // If the provider has no arguments (and therefore only generics),
@@ -206,15 +202,15 @@ ${parameters.map((e) => '        ${e.name}: ${e.name},\n').join()}
           : '() => ${provider.name}$typeParametersUsage()$cascadePropertyInit',
       parametersPassThrough: parametersPassThrough,
       other: '''
-abstract class $notifierTypedefName$typeParametersDefinition extends $notifierBaseType<${provider.valueType}> {
-  ${parameters.map((e) => 'late final ${e.type} ${e.name};').join('\n')}
+abstract class $notifierTypedefName$typeParametersDefinition extends $notifierBaseType<${provider.valueTypeDisplayString}> {
+  ${parameters.map((e) => 'late final ${e.typeDisplayString} ${e.name};').join('\n')}
 
-  ${provider.createdType} build($parameterDefinition);
+  ${provider.createdTypeDisplayString} build($parameterDefinition);
 }
 ''',
       providerOther: '''
   @override
-  ${provider.createdType} runNotifierBuild(
+  ${provider.createdTypeDisplayString} runNotifierBuild(
     covariant ${provider.name}$typeParametersUsage notifier,
   ) {
     return notifier.build($parametersPassThrough);
@@ -240,7 +236,7 @@ ${parameters.map((e) => '        ${e.name}: ${e.name},\n').join()}
   }
 
   final GeneratorProviderDeclaration provider;
-  final List<ParameterElement> parameters;
+  final List<FormalParameter> parameters;
   final TypeParameterList? typeParameters;
   final BuildYamlOptions options;
   final String refType;
@@ -346,7 +342,7 @@ class $providerTypeNameImpl$typeParametersDefinition extends $providerType$provi
     )}
   }) : super.internal();
 
-${parameters.map((e) => 'final ${e.type.getDisplayString(withNullability: true)} ${e.name};').join()}
+${parameters.map((e) => 'final ${e.typeDisplayString} ${e.name};').join()}
 
 $providerOther
 
@@ -378,18 +374,18 @@ ${[
   }
 }
 
-mixin $refNameImpl$typeParametersDefinition on $refType<${provider.valueType}> {
+mixin $refNameImpl$typeParametersDefinition on $refType<${provider.valueTypeDisplayString}> {
   ${parameters.map((e) {
       return '''
 /// The parameter `${e.name}` of this provider.
-${e.type} get ${e.name};''';
+${e.typeDisplayString} get ${e.name};''';
     }).join()}
 }
 
 class $elementNameImpl$typeParametersDefinition extends $elementType$providerGenerics with $refNameImpl$typeParametersUsage {
   $elementNameImpl(super.provider);
 
-${parameters.map((e) => '@override ${e.type} get ${e.name} => (origin as $providerTypeNameImpl$typeParametersUsage).${e.name};').join()}
+${parameters.map((e) => '@override ${e.typeDisplayString} get ${e.name} => (origin as $providerTypeNameImpl$typeParametersUsage).${e.name};').join()}
 }
 ''');
   }