+++ packages/riverpod_lint_flutter_test/test/lints/notifier_build/fix/notifier_build.dart
@@ -1,0 +1,8 @@
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+/// Fake Provider
+typedef _$ExampleProvider1 = Object;
+
+@riverpod
+// expect_lint: notifier_build
+class ExampleProvider1 extends _$ExampleProvider1 {}