+++ packages/hooks_riverpod/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.4.2 - 2023-10-02
+
+- `flutter_riverpod` upgraded to `2.4.2`
+
+## 2.4.1 - 2023-09-27
+
+- `flutter_riverpod` upgraded to `2.4.1`
+- `riverpod` upgraded to `2.4.1`
+
+## 2.4.0 - 2023-09-04
+
+- `flutter_riverpod` upgraded to `2.4.0`
+- `riverpod` upgraded to `2.4.0`
+
 ## 2.3.10 - 2023-08-28
 
 - `flutter_riverpod` upgraded to `2.3.10`