+++ packages/riverpod_analyzer_utils/CHANGELOG.md
@@ -1,6 +1,14 @@
-## Unreleased minor
+## Unreleased patch
 
 - Added `GeneratorProviderDeclarationElement.isFamily`
+
+## 0.3.5 - 2023-10-02
+
+- Correctly parse import aliases when used inside `ref.watch(<...>)`
+
+## 0.3.4 - 2023-09-27
+
+- Fixed `refInvocations` not getting parsed for generated providers with arguments.
 
 ## 0.3.3 - 2023-08-16