+++ packages/riverpod_lint/CHANGELOG.md
@@ -3,6 +3,23 @@
 - Updated `stateless_ref` and `generator_class_extends` to support providers
   with generic parameters.
 
+## 2.2.0 - 2023-10-02
+
+- Added `avoid_build_context_in_providers` (thanks to @charlescyt)
+- Fixed false positive with `avoid_manual_providers_as_generated_provider_dependency` when using import aliases
+
+## 2.1.1 - 2023-09-27
+
+- Fixed `provider_dependencies` lint causing false positives on providers with arguments.
+
+## 2.1.0 - 2023-09-14
+
+- Added `notifier_build`, a lint to catch when a Notifier has no `build` method (thansk to @LeonardoRosaa)
+
+## 2.0.4 - 2023-09-04
+
+- `riverpod` upgraded to `2.4.0`
+
 ## 2.0.3 - 2023-08-28
 
 - `riverpod` upgraded to `2.3.10`